Rename to ITrait
