+++ sources/Google.Solutions.IapDesktop.Core.Test/ClientModel/Protocol/TestAppProtocol.cs
@@ -21,7 +21,6 @@
 
 using Google.Solutions.IapDesktop.Core.ClientModel.Protocol;
 using Google.Solutions.IapDesktop.Core.ClientModel.Traits;
-using Google.Solutions.IapDesktop.Core.ClientModel.Transport;
 using Google.Solutions.IapDesktop.Core.ClientModel.Transport.Policies;
 using Google.Solutions.Testing.Apis;
 using Moq;
@@ -42,7 +41,7 @@ namespace Google.Solutions.IapDesktop.Core.Test.ClientModel.Protocol
         {
             return new AppProtocol(
                 name,
-                Enumerable.Empty<IProtocolTargetTrait>(),
+                Enumerable.Empty<ITrait>(),
                 new AllowAllPolicy(),
                 8080,
                 null,
@@ -71,14 +70,14 @@ namespace Google.Solutions.IapDesktop.Core.Test.ClientModel.Protocol
         {
             var protocol1 = new AppProtocol(
                 "app-1",
-                Enumerable.Empty<IProtocolTargetTrait>(),
+                Enumerable.Empty<ITrait>(),
                 new AllowAllPolicy(),
                 8080,
                 null,
                 null);
             var protocol2 = new AppProtocol(
                 "app-2",
-                Enumerable.Empty<IProtocolTargetTrait>(),
+                Enumerable.Empty<ITrait>(),
                 new AllowAllPolicy(),
                 8080,
                 null,
@@ -94,7 +93,7 @@ namespace Google.Solutions.IapDesktop.Core.Test.ClientModel.Protocol
         {
             var protocol1 = new AppProtocol(
                 "app-1",
-                Enumerable.Empty<IProtocolTargetTrait>(),
+                Enumerable.Empty<ITrait>(),
                 new AllowAllPolicy(),
                 8080,
                 null,
@@ -116,14 +115,14 @@ namespace Google.Solutions.IapDesktop.Core.Test.ClientModel.Protocol
         {
             var protocol1 = new AppProtocol(
                 "app-1",
-                Enumerable.Empty<IProtocolTargetTrait>(),
+                Enumerable.Empty<ITrait>(),
                 new AllowAllPolicy(),
                 8080,
                 null,
                 null);
             var protocol2 = new AppProtocol(
                 "app-1",
-                Enumerable.Empty<IProtocolTargetTrait>(),
+                Enumerable.Empty<ITrait>(),
                 new Mock<ITransportPolicy>().Object,
                 8080,
                 null,
@@ -139,14 +138,14 @@ namespace Google.Solutions.IapDesktop.Core.Test.ClientModel.Protocol
         {
             var protocol1 = new AppProtocol(
                 "app-1",
-                Enumerable.Empty<IProtocolTargetTrait>(),
+                Enumerable.Empty<ITrait>(),
                 new AllowAllPolicy(),
                 8080,
                 null,
                 null);
             var protocol2 = new AppProtocol(
                 "app-1",
-                Enumerable.Empty<IProtocolTargetTrait>(),
+                Enumerable.Empty<ITrait>(),
                 new AllowAllPolicy(),
                 80,
                 null,
@@ -162,14 +161,14 @@ namespace Google.Solutions.IapDesktop.Core.Test.ClientModel.Protocol
         {
             var protocol1 = new AppProtocol(
                 "app-1",
-                Enumerable.Empty<IProtocolTargetTrait>(),
+                Enumerable.Empty<ITrait>(),
                 new AllowAllPolicy(),
                 8080,
                 null,
                 null);
             var protocol2 = new AppProtocol(
                 "app-1",
-                Enumerable.Empty<IProtocolTargetTrait>(),
+                Enumerable.Empty<ITrait>(),
                 new AllowAllPolicy(),
                 8080,
                 null,
