+++ sources/Google.Solutions.IapDesktop.Core/ClientModel/Traits/TypeTrait.cs
@@ -26,7 +26,7 @@ using System.Linq;
 namespace Google.Solutions.IapDesktop.Core.ClientModel.Traits
 {
 	
-    public class InstanceTrait : IProtocolTargetTrait
+    public class InstanceTrait : ITrait
     {
         private const string Expression = "isInstance()";
 
@@ -52,7 +52,7 @@ namespace Google.Solutions.IapDesktop.Core.ClientModel.Traits
             return Equals(obj as InstanceTrait);
         }
 
-        public bool Equals(IProtocolTargetTrait other)
+        public bool Equals(ITrait other)
         {
             return other is InstanceTrait && other != null;
         }
@@ -105,7 +105,7 @@ namespace Google.Solutions.IapDesktop.Core.ClientModel.Traits
     }
 
 	
-    public class WindowsTrait : IProtocolTargetTrait
+    public class WindowsTrait : ITrait
     {
         private const string Expression = "isWindows()";
 
@@ -131,7 +131,7 @@ namespace Google.Solutions.IapDesktop.Core.ClientModel.Traits
             return Equals(obj as WindowsTrait);
         }
 
-        public bool Equals(IProtocolTargetTrait other)
+        public bool Equals(ITrait other)
         {
             return other is WindowsTrait && other != null;
         }
@@ -184,7 +184,7 @@ namespace Google.Solutions.IapDesktop.Core.ClientModel.Traits
     }
 
 	
-    public class LinuxTrait : IProtocolTargetTrait
+    public class LinuxTrait : ITrait
     {
         private const string Expression = "isLinux()";
 
@@ -210,7 +210,7 @@ namespace Google.Solutions.IapDesktop.Core.ClientModel.Traits
             return Equals(obj as LinuxTrait);
         }
 
-        public bool Equals(IProtocolTargetTrait other)
+        public bool Equals(ITrait other)
         {
             return other is LinuxTrait && other != null;
         }
